requirements.txt
alembic==1.5.3
asgiref==3.3.1
banal==1.0.1
Cerberus==1.3.2
certifi==2020.12.5
chardet==4.0.0
dataset==1.4.4
Django==3.1.7
django-cleanup==5.1.0
docopt==0.6.2
idna==2.10
Mako==1.1.4
MarkupSafe==1.1.1
mysql-connector-python==8.0.23
mysqlclient==2.0.3
peewee==3.14.0
peewee-mssql==0.1.3
pipdeptree==2.0.0
pipreqs==0.4.10
protobuf==3.14.0
pymssql==2.1.5
PyMySQL==1.0.2
pysqlite3==0.4.5
pysqlite3-binary==0.4.5
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.23
sqlit==0.1.6
sqlparse==0.4.1
urllib3==1.26.3
uuid==1.30
yarg==0.1.9
